+++ posthog/src/main/java/com/posthog/PostHog.kt
@@ -208,6 +208,7 @@ public class PostHog private constructor(
                 }
 
                 // only because of testing in isolation, this flag is always enabled
+                @Suppress("DEPRECATION")
                 if (reloadFeatureFlags) {
                     when {
                         config.remoteConfig ->