+++ posthog/src/test/java/com/posthog/PostHogBeforeSendTest.kt
@@ -35,6 +35,7 @@ internal class PostHogBeforeSendTest {
         val trigger: (PostHogInterface) -> Unit,
     )
 
+    @Suppress("DEPRECATION")
     fun getSut(
         host: String,
         flushAt: Int = 1,