+++ posthog/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- chore: Remove `config=true` from flags endpoint, add `timezone` to flags requests, deprecate `remoteConfig` option ([#421](https://github.com/PostHog/posthog-android/pull/421))
+
 ## 6.6.3
 
 ### Patch Changes